Large int bug (#238)
There was an issue where in some environments (but not all) large numbers such as 8.034023767017109e+27 whch were ints would be printed as an the int 8034023767017108950029959168 and then the C++ compiler would complain as that's more than the maximum int value. This has been fixed by only printing ints as ints if they are MIN_INT < number < MAX_INT else we print them as a float (and we get the sceintific notation).